LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-blur-and-focus-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="container">
<input id="before">
<input id="test" type="time" value="12:34">
<input id="after">
</div>
<script>
description('Check blur and focus events for multiple fields time input UI');

var blurCounter = 0;
var focusCounter = 0;

function keyDown(key, modifiers)
{
    if (!window.eventSender)
        return;
    eventSender.keyDown(key, modifiers);
}

function state()
{
    return 'blur=' + blurCounter + ' focus=' + focusCounter;
}

var testInput = document.getElementById('test');
testInput.addEventListener('blur', function () { ++blurCounter; });
testInput.addEventListener('focus', function () { ++focusCounter; });

function beginTestCase(testCaseName)
{
    debug(testCaseName);
    blurCounter = 0;
    focusCounter = 0;
}

beginTestCase('focus() and blur()');
shouldBeEqualToString('testInput.focus(); state()', 'blur=0 focus=1');
shouldBeEqualToString('testInput.blur(); state()', 'blur=1 focus=1');

if (window.eventSender) {
    beginTestCase('focus and Tab key to blur');
    document.getElementById("before").focus();
    shouldBeEqualToString('keyDown("\t"); state()', 'blur=0 focus=1');
    shouldBeEqualToString('keyDown("\t"); state()', 'blur=0 focus=1');
    shouldBeEqualToString('keyDown("\t"); state()', 'blur=0 focus=1');
    shouldBeEqualToString('keyDown("\t"); state()', 'blur=1 focus=1');
} else {
    debug('Please run in DumpRenderTree for focus and Tab-key test case');
}

debug('');
document.body.removeChild(document.getElementById("container"));
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
